ui/login: rename file
